dictation/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11
gTTS==1.2.2
gTTS-token==1.1.1
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
google-cloud-storage==1.6.0